+++ tests/integration/src/rust_masm_tests/mod.rs
@@ -4,9 +4,11 @@
 use std::{collections::VecDeque, sync::Arc};
 
 use miden_core::Felt;
-use miden_debug::{Executor, FromMidenRepr};
+use miden_debug::{Executor, FromMidenRepr, ToMidenRepr};
 use midenc_session::Session;
 use proptest::{prop_assert_eq, test_runner::TestCaseError};
+
+use crate::testing::eval_package;
 
 mod abi_transform;
 mod apps;
@@ -26,13 +28,13 @@ pub fn run_masm_vs_rust<T>(
 where
     T: Clone + FromMidenRepr + PartialEq + std::fmt::Debug,
 {
-    let exec = Executor::for_package(package, args.iter().copied())
-        .map_err(|err| TestCaseError::fail(err.to_string()))?;
-    let output = exec.execute_into(&package.unwrap_program(), session.source_manager.clone());
-    std::dbg!(&output);
-    prop_assert_eq!(rust_out.clone(), output, "VM output mismatch");
-    // TODO: Uncomment after https://github.com/0xMiden/compiler/issues/228 is fixed
-    // let emul_out: T = (*execute_emulator(ir_program.clone(), args).first().unwrap()).into();
-    // prop_assert_eq!(rust_out, emul_out, "Emulator output mismatch");
+    eval_package::<Felt, _, _>(package, None, args, session, |trace| {
+        let vm_out_felt0 = trace.outputs().get_stack_item(0).unwrap();
+        let vm_out_felt1 = trace.outputs().get_stack_item(1).unwrap();
+        let vm_out: T = T::from_felts(&[vm_out_felt0, vm_out_felt1]);
+        dbg!(&vm_out);
+        prop_assert_eq!(rust_out.clone(), vm_out, "VM output mismatch");
+        Ok(())
+    })?;
     Ok(())
 }